Keep chunk reconstruction test-only
