+++ packages/core/src/rdf-literal-size.ts
@@ -76,19 +76,6 @@ export interface RdfLiteralNormalizationOptions {
   readonly label?: string;
 }
 
-export interface ReconstructedChunkedTextBody {
-  readonly subject: string;
-  readonly bodySubject: string;
-  readonly sourcePredicate: string;
-  readonly lexical: string;
-  readonly literalTerm: string;
-  readonly chunkCount: number;
-  readonly lexicalSha256: string;
-  readonly literalTermSha256: string;
-  readonly language?: string;
-  readonly datatype?: string;
-}
-
 export class OversizedRdfLiteralError extends DKGUserError {
   readonly code = OVERSIZED_RDF_LITERAL_ERROR_CODE;
   readonly actualBytes: number;
@@ -319,88 +306,6 @@ export function normalizeLargeRdfLiteralsForBlazegraph(
   return { quads: normalized, rewrites };
 }
 
-export function reconstructChunkedTextBodies(
-  quads: readonly QuadLiteralLike[],
-  options: { subject?: string; bodySubject?: string; sourcePredicate?: string } = {},
-): ReconstructedChunkedTextBody[] {
-  const bySubject = indexQuadsBySubject(quads);
-  const bodyLinks = quads.filter((q) =>
-    q.predicate === DKG_HAS_TEXT_BODY &&
-    (!options.subject || q.subject === options.subject) &&
-    (!options.bodySubject || q.object === options.bodySubject)
-  );
-  const explicitBody = options.bodySubject && bodyLinks.length === 0
-    ? [{ subject: findOwnerSubject(quads, options.bodySubject), object: options.bodySubject }]
-    : [];
-  const bodies = [...bodyLinks, ...explicitBody];
-  const reconstructed: ReconstructedChunkedTextBody[] = [];
-
-  for (const link of bodies) {
-    const bodySubject = link.object;
-    const bodyQuads = bySubject.get(bodySubject) ?? [];
-    const sourcePredicate = iriObject(bodyQuads, DKG_TEXT_SOURCE_PREDICATE);
-    if (!sourcePredicate) throw new Error(`Chunked text body ${bodySubject} is missing source predicate`);
-    if (options.sourcePredicate && sourcePredicate !== options.sourcePredicate) {
-      continue;
-    }
-
-    const count = integerObject(bodyQuads, DKG_TEXT_CHUNK_COUNT);
-    if (count === undefined) throw new Error(`Chunked text body ${bodySubject} is missing chunk count`);
-    const lexicalSha256 = literalObject(bodyQuads, DKG_TEXT_CONTENT_SHA256);
-    if (!lexicalSha256) throw new Error(`Chunked text body ${bodySubject} is missing content hash`);
-    const literalTermSha256 = literalObject(bodyQuads, DKG_TEXT_LITERAL_TERM_SHA256);
-    if (!literalTermSha256) throw new Error(`Chunked text body ${bodySubject} is missing literal term hash`);
-    const language = literalObject(bodyQuads, DKG_TEXT_LANGUAGE);
-    const datatype = iriObject(bodyQuads, DKG_TEXT_DATATYPE);
-    const chunkSubjects = bodyQuads.filter((q) => q.predicate === DKG_HAS_TEXT_CHUNK).map((q) => q.object);
-    if (chunkSubjects.length !== count) {
-      throw new Error(`Chunked text body ${bodySubject} expected ${count} chunks but found ${chunkSubjects.length}`);
-    }
-
-    const chunks = chunkSubjects.map((chunkSubject) => {
-      const chunkQuads = bySubject.get(chunkSubject) ?? [];
-      const index = integerObject(chunkQuads, DKG_CHUNK_INDEX);
-      if (index === undefined) throw new Error(`Chunk ${chunkSubject} is missing chunkIndex`);
-      const valueTerm = literalTermObject(chunkQuads, DKG_CHUNK_VALUE);
-      if (!valueTerm) throw new Error(`Chunk ${chunkSubject} is missing chunkValue`);
-      const parsed = parseRdfLiteralTerm(valueTerm);
-      if (!parsed) throw new Error(`Chunk ${chunkSubject} has invalid chunkValue`);
-      return { index, lexical: parsed.lexical };
-    }).sort((a, b) => a.index - b.index);
-
-    for (let i = 0; i < chunks.length; i++) {
-      if (chunks[i]!.index !== i) {
-        throw new Error(`Chunked text body ${bodySubject} has non-contiguous chunk index ${chunks[i]!.index}`);
-      }
-    }
-
-    const lexical = chunks.map((chunk) => chunk.lexical).join('');
-    const suffix = suffixFromMetadata(language, datatype);
-    const literalTerm = rdfLiteralTerm(lexical, suffix);
-    if (sha256Hex(lexical) !== lexicalSha256) {
-      throw new Error(`Chunked text body ${bodySubject} content hash mismatch`);
-    }
-    if (sha256Hex(literalTerm) !== literalTermSha256) {
-      throw new Error(`Chunked text body ${bodySubject} literal term hash mismatch`);
-    }
-
-    reconstructed.push({
-      subject: link.subject,
-      bodySubject,
-      sourcePredicate,
-      lexical,
-      literalTerm,
-      chunkCount: chunks.length,
-      lexicalSha256,
-      literalTermSha256,
-      ...(language !== undefined ? { language } : {}),
-      ...(datatype !== undefined ? { datatype } : {}),
-    });
-  }
-
-  return reconstructed;
-}
-
 function validateNormalizationLimits(maxBytes: number, chunkMaxBytes: number): void {
   if (!Number.isInteger(maxBytes) || maxBytes <= 0 || maxBytes > JAVA_WRITE_UTF_MAX_BYTES) {
     throw new Error(`Invalid maxBytes: ${maxBytes}`);
@@ -595,47 +500,4 @@ function sha256Hex(value: string): string {
 
 function utf8ByteLength(value: string): number {
   return new TextEncoder().encode(value).length;
-}
-
-function indexQuadsBySubject(quads: readonly QuadLiteralLike[]): Map<string, QuadLiteralLike[]> {
-  const map = new Map<string, QuadLiteralLike[]>();
-  for (const quad of quads) {
-    const list = map.get(quad.subject);
-    if (list) list.push(quad);
-    else map.set(quad.subject, [quad]);
-  }
-  return map;
-}
-
-function findOwnerSubject(quads: readonly QuadLiteralLike[], bodySubject: string): string {
-  return quads.find((q) => q.predicate === DKG_HAS_TEXT_BODY && q.object === bodySubject)?.subject ?? '';
-}
-
-function literalTermObject(quads: readonly QuadLiteralLike[], predicate: string): string | undefined {
-  const object = quads.find((q) => q.predicate === predicate)?.object;
-  return object?.startsWith('"') ? object : undefined;
-}
-
-function literalObject(quads: readonly QuadLiteralLike[], predicate: string): string | undefined {
-  const term = literalTermObject(quads, predicate);
-  if (!term) return undefined;
-  const parsed = parseRdfLiteralTerm(term);
-  return parsed?.lexical;
-}
-
-function iriObject(quads: readonly QuadLiteralLike[], predicate: string): string | undefined {
-  const object = quads.find((q) => q.predicate === predicate)?.object;
-  return object && !object.startsWith('"') ? object : undefined;
-}
-
-function integerObject(quads: readonly QuadLiteralLike[], predicate: string): number | undefined {
-  const value = literalObject(quads, predicate);
-  if (value === undefined || !/^\d+$/.test(value)) return undefined;
-  return Number(value);
-}
-
-function suffixFromMetadata(language?: string, datatype?: string): string {
-  if (language) return `@${language}`;
-  if (datatype) return `^^<${datatype}>`;
-  return '';
 }
